'audioVolume' is not handled correctly by is_setting_supported
The 'audioVolume' setting has a very specific definition in the DAB specification, for system/settings/list, unlike the other settings who are either lists/arrays or booleans.
'audioVolume' is defined as a dictionary, with 2 elements 'min' and 'max'. Moreover, the volume control is considered disabled if 'min' and 'max' have the same value.
Currently, is_setting_supported() doesn't handle 'audioVolume' correctly, because it can handle only lists or booleans.
So a specific handling for 'audioVolume' is added, to match the DAB specification.